Compatibility with Xcode 7.3 and iOS9.3
Linked files in class folder as they were shown as missing. Added Zip.h
and Zip.m files into the project. They existing in the classes folder,
but were not linked in the project. Changed the Titanium SDK path in
the Titanium.xcconfig file. Updated the Titanium SDK version to 5.2.2.
Added updated lib.1.2.8.dylib. The 1.2.5 version was missing from the
project.